Added new source files for each view that was previously included in the ContentView source file.
