+++ projectSkeleton/ContentView.swift
@@ -441,212 +441,6 @@ struct ContentView: View
     } // NavigationView
 }
 
-/*
-struct MultipleSelectionList: View {
-    @State var items: [String] = ["Apples", "Oranges", "Bananas", "Pears", "Mangos", "Grapefruit"]
-    @State var selections: [String] = []
-
-    var body: some View {
-        List {
-            ForEach(self.items, id: \.self) { item in
-                MultipleSelectionRow(title: item, isSelected: self.selections.contains(item)) {
-                    if self.selections.contains(item) {
-                        self.selections.removeAll(where: { $0 == item })
-                    }
-                    else {
-                        self.selections.append(item)
-                    }
-                }
-            }
-        }
-    }
-}
- 
- struct MultipleSelectionRow: View {
-     var title: String
-     var isSelected: Bool
-     var action: () -> Void
-
-     var body: some View {
-         Button(action: self.action) {
-             HStack {
-                 Text(self.title)
-                 if self.isSelected {
-                     Spacer()
-                     Image(systemName: "checkmark")
-                 }
-             }
-         }
-     }
- }
- */
-
-struct MultipleSelectionRow: View
-{
-    var name : String
-    var isSelected : Bool
-    var action: () -> Void
-
-    var body: some View
-    {
-        Button(action: self.action)
-        {
-            HStack
-            {
-                Text(self.name)
-                Spacer()
-                if self.isSelected
-                {
-                    Image("GreenCircle")
-                    .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 20, maxHeight: 20)
-                }
-                else
-                {
-                    Image("GrayCircle")
-                    .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 20, maxHeight: 20)
-                }
-            }
-        }
-    }
-}
-
-struct AddListView: View
-{
-    @Environment(\.managedObjectContext) var objContext
-    @FetchRequest(entity: Satellite.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \Satellite.name, ascending: true)]) var satellites : FetchedResults<Satellite>
-    @Binding var inList : [SatelliteSelect]
-    @State var selectedList : [SatelliteSelect] = []
-    
-    var body: some View
-    {
-        List
-        {
-            ForEach(inList, id: \.self)
-            {
-                satellite in
-                MultipleSelectionRow(name: satellite.name, isSelected:  selectedList.contains(satellite))
-                {
-                    if selectedList.contains(satellite)
-                    {
-                        selectedList.removeAll(where: { $0 == satellite })
-                    }
-                    else
-                    {
-                        selectedList.append(satellite)
-                    }
-                }
-            }
-        }
-        .onAppear(perform: {selectedList = []})
-        .listStyle(GroupedListStyle())
-        .navigationBarTitle("Select Satellites to Add", displayMode: .inline)
-        .navigationBarItems(trailing:
-            Button(action:
-            {
-                for selected in selectedList
-                {
-                    let isUnique : Bool = ((satellites.first(where: {$0.wrappedName == selected.name})) == nil)
-                    if (isUnique)
-                    {
-                        var newSatellite : Satellite = Satellite(context: objContext)
-                        newSatellite.name = selected.name
-                        newSatellite.tleLine1 = selected.tle1
-                        newSatellite.tleLine2 = selected.tle2
-                        newSatellite.age = Date.now
-                        
-                        try? objContext.save()
-                    }
-                }
-            })
-            {
-                Text("OK")
-            }
-        )
-    }
-}
-    
-struct ListView: View
-{
-    @Environment(\.managedObjectContext) var objContext
-    @FetchRequest(entity: Satellite.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \Satellite.name, ascending: true)]) var satellites : FetchedResults<Satellite>
-    @Binding var time : Date
-    
-    var body: some View
-    {
-        List
-        {
-            Text("Satellite Search").font(.system(size: 30))
-            
-            ForEach(satellites)
-            {
-                satellite in
-                let coords : GeoCoords = calculateGeoCoords(inSatellite: satellite, inTime: time)
-                VStack
-                {
-                    HStack
-                    {
-                        VStack
-                        {
-                            Text(satellite.wrappedName)
-                            Text(String(coords.longitude))
-                            Text(String(coords.latitude))
-                            Text(String(coords.altitude))
-                        }
-                        if (satellite.isFavorite)
-                        {
-                            Image("YellowStar")
-                                .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 50, maxHeight: 50)                    .onTapGesture
-                            {
-                                satellite.isFavorite.toggle()
-                                
-                                do
-                                {
-                                    try objContext.save()
-                                }
-                                catch
-                                {
-                                    // error
-                                }
-                            }
-                        }
-                        else
-                        {
-                            Image("GrayStar")
-                                .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 50, maxHeight: 50)                    .onTapGesture
-                            {
-                                satellite.isFavorite.toggle()
-                                
-                                do
-                                {
-                                    try objContext.save()
-                                }
-                                catch
-                                {
-                                    // error
-                                }
-                            }
-                        }
-                        Image("Trash")
-                            .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 50, maxHeight: 50)                    .onTapGesture
-                        {
-                            objContext.delete(satellite)
-                            do
-                            {
-                                try objContext.save()
-                            }
-                            catch
-                            {
-                                // error
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
 struct ContentView_Previews: PreviewProvider
 {
     static var previews: some View
@@ -654,101 +448,3 @@ struct ContentView_Previews: PreviewProvider
         ContentView()
     }
 }
-    
-    
-    /*
-     // Codable: Allows the object to be decoded from one representation and encoded into another.
-     struct OrbitalDataJSON : Codable
-     {
-     var OBJECT_NAME : String = ""
-     var OBJECT_ID : String = ""
-     var EPOCH : String = ""
-     var MEAN_MOTION : Double = 0
-     var ECCENTRICITY : Double = 0
-     var INCLINATION : Double = 0
-     var RA_OF_ASC_NODE : Double = 0
-     var ARG_OF_PERICENTER : Double = 0
-     var MEAN_ANOMALY : Double = 0
-     var EPHEMERIS_TYPE : Int = 0
-     var CLASSIFICATION_TYPE : String = ""
-     var NORAD_CAT_ID : Int = 0
-     var ELEMENT_SET_NO : Int = 0
-     var REV_AT_EPOCH : Int = 0
-     var BSTAR : Double = 0
-     var MEAN_MOTION_DOT : Double = 0
-     var MEAN_MOTION_DDOT : Double = 0
-     }
-     */
-    
-    /*
-     func getSatellite2()
-     {
-     let urlAsString : String = "https://celestrak.org/NORAD/elements/gp.php?" + gQuery + gValue + "&FORMAT=JSON-PRETTY"
-     let testURL : String = "https://celestrak.org/NORAD/elements/gp.php?NAME=MICROSAT-R&FORMAT=JSON-PRETTY"
-     let url = URL(string: urlAsString)!
-     let urlSession = URLSession.shared
-     
-     // jsonQuery is a new task that will run separately to the main thread
-     let jsonQuery = urlSession.dataTask(with: url, completionHandler:
-     {
-     data, response, error -> Void in
-     if (error != nil)
-     {
-     print(error!.localizedDescription)
-     }
-     
-     do {
-     let decoder = JSONDecoder()
-     // TRICKY: Note the use of [OrbitalData] result as an array
-     // as the Celestrak REST API returns JSON results in a JSON array
-     let jsonResult = try decoder.decode([OrbitalDataJSON].self, from: data!)
-     if (jsonResult != nil)
-     {
-     var orbitalData : OrbitalDataJSON = jsonResult[0]
-     var newSatellite : Satellite = Satellite()
-     
-     newSatellite.mName = orbitalData.OBJECT_NAME
-     newSatellite.mData2 = orbitalData
-     
-     // TRICKY: Calling a C routine that needs to return values using C pointers
-     // In SwiftUI, we simulate C pointers using UnsafeMutablePointer<>
-     var age : Double = 0
-     var ageP : UnsafeMutablePointer<Double> = .init(&age)
-     var lat : Double = 0
-     var latP : UnsafeMutablePointer<Double> = .init(&lat)
-     var lon : Double = 0
-     var lonP : UnsafeMutablePointer<Double> = .init(&lon)
-     var alt : Double = 0
-     var altP : UnsafeMutablePointer<Double> = .init(&alt)
-     
-     // TRICKY: Call C function here using the SwiftUI binding header decl.
-     orbit_to_lla(nil, nil, nil, ageP, latP, lonP, altP)
-     newSatellite.mLatitude = lat
-     newSatellite.mLongitude = lon
-     
-     var isUnique : Bool = ((satellites.first(where: {$0.mName == newSatellite.mName})) == nil)
-     if (isUnique)
-     {
-     satellites.append(newSatellite)
-     }
-     }
-     } catch DecodingError.dataCorrupted(let context) {
-     print(context)
-     } catch DecodingError.keyNotFound(let key, let context) {
-     print("Key '\(key)' not found:", context.debugDescription)
-     print("codingPath:", context.codingPath)
-     } catch DecodingError.valueNotFound(let value, let context) {
-     print("Value '\(value)' not found:", context.debugDescription)
-     print("codingPath:", context.codingPath)
-     } catch DecodingError.typeMismatch(let type, let context) {
-     print("Type '\(type)' mismatch:", context.debugDescription)
-     print("codingPath:", context.codingPath)
-     } catch {
-     print("error: ", error)
-     }
-     })
-     // resume() tells the task to start running on its own thread
-     jsonQuery.resume()
-     }
-     */
-
