+++ projectSkeleton/MultipleSelectionRowView.swift
@@ -1,0 +1,39 @@
+//
+//  MultipleSelectionRowView.swift
+//  projectSkeleton
+//
+//  Created by Matthew Fitzgerald on 1/14/23.
+//
+import CoreData
+import SwiftUI
+import Foundation
+import MapKit
+
+struct MultipleSelectionRow: View
+{
+    var name : String
+    var isSelected : Bool
+    var action: () -> Void
+
+    var body: some View
+    {
+        Button(action: self.action)
+        {
+            HStack
+            {
+                Text(self.name)
+                Spacer()
+                if self.isSelected
+                {
+                    Image("GreenCircle")
+                    .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 20, maxHeight: 20)
+                }
+                else
+                {
+                    Image("GrayCircle")
+                    .resizable().aspectRatio(contentMode: .fit).frame(maxWidth: 20, maxHeight: 20)
+                }
+            }
+        }
+    }
+}
